chore: rename util/is-proxy-disabled to util/network-mode
The module has exported isBrowserNetworkMode since the forceHttp1 rework
and its debug tag already says cypress:server:network-mode. A pure rename
keeps the v8 snapshot classification, so the snapshot-meta entries move
in place.
