+++ tooling/v8-snapshot/cache/win32/snapshot-meta.json
@@ -3743,8 +3743,8 @@
     "./packages/server/lib/util/graceful_crash_handling.ts",
     "./packages/server/lib/util/headers.ts",
     "./packages/server/lib/util/human_time.ts",
-    "./packages/server/lib/util/is-proxy-disabled.ts",
     "./packages/server/lib/util/net_profiler.ts",
+    "./packages/server/lib/util/network-mode.ts",
     "./packages/server/lib/util/network_failures.ts",
     "./packages/server/lib/util/newlines.ts",
     "./packages/server/lib/util/patch-fs.ts",
